+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - 8e3130d Added scrutinizer check for code style and coverage
 - dbaf84f Added Mollie standard types
 - 1b53ea9-payments resource implementation
+- f658c63-fixed go linting errors
+- 8893236-methods resource implementation
+
 
 ### Deprecated
 - Nothing